Reorganized test of segments
