+++ src/fastoad/models/performances/mission/segments/tests/conftest.py
@@ -1,0 +1,63 @@
+#  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
+#  Copyright (C) 2022 ONERA & ISAE-SUPAERO
+#  FAST is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#  You should have received a copy of the GNU General Public License
+#  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+import numpy as np
+import pandas as pd
+import pytest
+
+from fastoad.model_base import FlightPoint
+from fastoad.model_base.propulsion import AbstractFuelPropulsion
+from ...polar import Polar
+
+
+@pytest.fixture
+def polar() -> Polar:
+    """Returns a dummy polar where max L/D ratio is around 16."""
+    cl = np.arange(0.0, 1.5, 0.01)
+    cd = 0.5e-1 * cl ** 2 + 0.01
+    return Polar(cl, cd)
+
+
+def print_dataframe(df):
+    """Utility for correctly printing results"""
+    # Not used if all goes well. Please keep it for future test setting/debugging.
+    with pd.option_context(
+        "display.max_rows", 20, "display.max_columns", None, "display.width", None
+    ):
+        print()
+        print(df)
+
+
+class DummyEngine(AbstractFuelPropulsion):
+    def __init__(self, max_thrust, max_sfc):
+        """
+        Dummy engine model.
+
+        Max thrust does not depend on flight conditions.
+        SFC varies linearly with thrust_rate, from max_sfc/2. at thrust rate is 0.,
+        to max_sfc when thrust_rate is 1.0
+
+        :param max_thrust: thrust when thrust rate = 1.0
+        :param max_sfc: SFC when thrust rate = 1.0
+        """
+        self.max_thrust = max_thrust
+        self.max_sfc = max_sfc
+
+    def compute_flight_points(self, flight_point: FlightPoint):
+
+        if flight_point.thrust_is_regulated or flight_point.thrust_rate is None:
+            flight_point.thrust_rate = flight_point.thrust / self.max_thrust
+        else:
+            flight_point.thrust = self.max_thrust * flight_point.thrust_rate
+
+        flight_point.sfc = self.max_sfc * (1.0 + flight_point.thrust_rate) / 2.0
